# Backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pigeonchat API",
        "version": "1.0.0",
        "description": "API for pigeonchat application",
        "contact": {
            "name": "Vivek Maniyar",
            "email": "vivekmaniyar9601@gmail.com"
        }
    },
    "paths":{
...
        "/users?search={search}":{
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Search for a user",
                "description": "Search for a user",
                "operationId": "searchUser",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Search string",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/searchresult"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/messages/{chatId}":{
            "get":{
                "tags":[
                    "Messages"
                ],
                "summary":"Get messages of a chat",
                "description":"Get messages of a chat",
                "operationId":"getMessages",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID of chat to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components":{
        "schemas":{
            "error":{
                "type":"object",
                "properties":{
                    "name":{
                        "type":"string"
                    },
                    "message":{
                        "type":"string"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties":{
                    "name": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "profileimage": {
                        "type": "string"
                    }
                }
            },
            "Message":{
                "type":"object",
                "properties":{
                    "message":{
                        "type":"string"
                    },
                    "sender":{
                        "$ref":"#/components/schemas/User"
                    },
                    "chat":{
                        "$ref":"#/components/schemas/Chat"
                    },
                    "readby":{
                        "type":"array",
                        "items":{
                            "$ref":"#/components/schemas/User"
                        }
                    }
                }
            },
            "Chat":{
                "type":"object",
                "properties":{
                    "chatName":{
                        "type":"string"
                    },
                    "isGroupChat":{
                        "type":"boolean"
                    },
                    "users":{
                        "type":"array",
                        "items":{
                            "$ref":"#/components/schemas/User"
                        }
                    },
                    "latestMessage":{
                        "$ref":"#/components/schemas/Message"
                    },
                    "groupadmin":{
                        "$ref":"#/components/schemas/User"
                    }
                }
            },
...
            "searchresult":{
                "type":"object",
                "properties":{
                    "_id":{
                        "type":"string"
                    },
                    "name":{
                        "type":"string"
                    },
                    "username":{
                        "type":"string"
                    },
                    "email":{
                        "type":"string"
                    },
                    "profileimage":{
                        "type":"string"
                    }
                    
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    }
}